backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langgraph
cohere
chromadb
pypdf
langchain-text-splitters
sqlalchemy
pytest
passlib[bcrypt]